Install simply Passenger package instead of running a role.
The task "Restart apache" was removed because there is no need for
restart in case of database or redmine file synchronization. However,
one needs to reload Apache whenever redmine_config changes and for
that we notify the handler "reload apache" in apache_vhost-role. The
same handler is used also for the passenger package installation.